% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findWeekBoundaries.R
\name{findWeekBoundaries}
\alias{findWeekBoundaries}
\title{Given a set of dates, it finds the week numbers they belong to and the initial and
last date for each week}
\usage{
findWeekBoundaries(dates)
}
\arguments{
\item{dates}{}
}
\description{
Given a set of dates, it finds the week numbers they belong to and the initial and
last date for each week
}
