% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessTeam.R
\name{guessTeam}
\alias{guessTeam}
\title{Guesses the team of a player based on the most frequent team on the games for each season}
\usage{
guessTeam(appearances_wo_teams)
}
\arguments{
\item{appearances_wo_teams}{}
}
\description{
Guesses the team of a player based on the most frequent team on the games for each season
}
